// src/app/interfaces/caretelera.interface.ts
// Generated by https://quicktype.io esta aplicacon para generar interfaces de TypeScript a partir de JSON
// Para generar la interface desde la respuesta de la API de TMDB
// Usamos la extensión "Paste JSON as Code" para que VS Code genere la interface automáticamente
// Para pegar es Ctrl + Shift + V
// Se pega el JSON de la respuesta de la API de TMDB

// Interface principal que representa la respuesta de la cartelera de películas
export interface CarteleraResponse {
  dates: Dates; // Rango de fechas de la cartelera
  page: number; // Página actual de la respuesta
  results: Movie[]; // Lista de películas
  total_pages: number; // Total de páginas disponibles
  total_results: number; // Total de resultados encontrados
}

// Interface para el rango de fechas de la cartelera
export interface Dates {
  maximum: string; // Fecha máxima (última)
  minimum: string; // Fecha mínima (primera)
}

// Interface que representa una película individual
export interface Movie {
  adult: boolean; // Indica si la película es para adultos
  backdrop_path: string; // Ruta de la imagen de fondo
  genre_ids: number[]; // IDs de los géneros de la película
  id: number; // ID único de la película
  original_language: OriginalLanguage; // Idioma original
  original_title: string; // Título original
  overview: string; // Sinopsis de la película
  popularity: number; // Popularidad
  poster_path: string; // Ruta del póster
  release_date: string; // Fecha de estreno
  title: string; // Título de la película
  video: boolean; // Indica si tiene video
  vote_average: number; // Promedio de votos
  vote_count: number; // Cantidad de votos
}

// Enum para los posibles idiomas originales de la película
export enum OriginalLanguage {
  En = 'en', // Inglés
  Es = 'es', // Español
  Pt = 'pt', // Portugués
}